fix(runtime-vapor): unit test error
