+++ packages/runtime-vapor/src/errorHandling.ts
@@ -3,10 +3,10 @@
 // The ultimate aim is to uncouple this replicated code and
 // facilitate its shared use between two runtimes.
 
-import { VaporLifecycleHooks } from './apiLifecycle'
 import { type ComponentInternalInstance } from './component'
 import { isFunction, isPromise } from '@vue/shared'
 import { warn } from './warning'
+import { VaporLifecycleHooks } from './enums'
 
 // contexts where user provided function may be executed, in addition to
 // lifecycle hooks.
